although this old relic of a script was written in the year 2011, it has been running weekly without fail ever since. this script is like a proto-finick :) and it has proven to be useful and effective.
this script is used to generate a weekly email (or monthly email, you choose time frame) that assigns each team-member some commits authored by others, so that they can be code-reviewed. the volume of lines to be reviewed each time is configurable. commits are chosen for review at random. multiple repositories can be handled in one weekly email, as can multiple branches of the same repository.